+++ AltinnApp/AT.Common.AltinnApp.Test/Unit/StructuredDataManagerTests.cs
@@ -4,6 +4,7 @@ using Altinn.App.Core.Internal.Data;
 using Altinn.Platform.Storage.Interface.Models;
 using Arbeidstilsynet.Common.AltinnApp.Extensions;
 using Arbeidstilsynet.Common.AltinnApp.Implementation;
+using Arbeidstilsynet.Common.AltinnApp.Test.Unit.TestFixtures;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using Shouldly;
@@ -44,8 +45,10 @@ public class StructuredDataManagerTests
     public async Task End_TaskEnd_UsesIApplicationClientCorrectly()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
         var dataModel = new TestDataModel { Name = "Test" };
 
         _applicationClient.GetApplication(instance.Org, "testApp").Returns(application);
@@ -68,8 +71,10 @@ public class StructuredDataManagerTests
     public async Task End_TaskEnd_UsesIDataClientToGetFormData()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
         var dataModel = new TestDataModel { Name = "Test" };
         var expectedGuid = Guid.Parse(instance.Data.First().Id);
 
@@ -101,8 +106,10 @@ public class StructuredDataManagerTests
     public async Task End_TaskEnd_UsesIDataClientToInsertStructuredData()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
         var dataModel = new TestDataModel { Name = "Test" };
 
         _applicationClient
@@ -136,8 +143,10 @@ public class StructuredDataManagerTests
     public async Task End_ProcessEnd_UsesIApplicationClientToGetDataElement()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
 
         _applicationClient.GetApplication(instance.Org, "testApp").Returns(application);
 
@@ -152,8 +161,10 @@ public class StructuredDataManagerTests
     public async Task End_ProcessEnd_UsesIDataClientToDeleteData()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
         var dataElement = instance.Data.First();
         var expectedGuid = Guid.Parse(dataElement.Id);
 
@@ -180,8 +191,14 @@ public class StructuredDataManagerTests
     public async Task End_ProcessEnd_When_DeleteAppDataModelAfterMapping_IsFalse_ShouldNotDeleteData()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
+
+        _applicationClient
+            .GetApplication(Arg.Any<string>(), Arg.Any<string>())
+            .Returns(application);
 
         var sut_withDeleteDisabled = new StructuredDataManager<TestDataModel, TestStructuredData>(
             _applicationClient,
@@ -192,10 +209,6 @@ public class StructuredDataManagerTests
             },
             _logger
         );
-
-        _applicationClient
-            .GetApplication(Arg.Any<string>(), Arg.Any<string>())
-            .Returns(application);
 
         // Act
         await sut_withDeleteDisabled.End(instance, new List<InstanceEvent>());
@@ -216,8 +229,10 @@ public class StructuredDataManagerTests
     public async Task End_ProcessEnd_RemovesDataElementFromInstance()
     {
         // Arrange
-        var instance = CreateTestInstance();
-        var application = CreateTestApplication();
+        var instance = AltinnData.CreateTestInstance();
+        var application = AltinnData.CreateTestApplication(
+            classRef: typeof(TestDataModel).FullName
+        );
         var initialDataCount = instance.Data.Count;
 
         _applicationClient
@@ -229,36 +244,6 @@ public class StructuredDataManagerTests
 
         // Assert
         instance.Data.Count.ShouldBe(initialDataCount - 1);
-    }
-
-    private static Instance CreateTestInstance()
-    {
-        return new Instance
-        {
-            Id = "50001234/fa0678ad-960d-4307-aba2-ba29c9804c9d",
-            Org = "testOrg",
-            AppId = "testOrg/testApp",
-            InstanceOwner = new InstanceOwner { PartyId = "50001234" },
-            Data = new List<DataElement>
-            {
-                new() { Id = "12345678-1234-1234-1234-123456789012", DataType = "testDataType" },
-            },
-        };
-    }
-
-    private static Application CreateTestApplication()
-    {
-        return new Application
-        {
-            DataTypes = new List<DataType>
-            {
-                new()
-                {
-                    Id = "testDataType",
-                    AppLogic = new ApplicationLogic() { ClassRef = typeof(TestDataModel).FullName },
-                },
-            },
-        };
     }
 
     public class TestDataModel